requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.1
plotly==5.9.0
seaborn==0.13.0
streamlit==1.26.0
streamlit_option_menu==0.3.6
